[dev] Wrote faster list and updated docs
Implemented a sparse list data structure to speed up the MEM. Since we
aren't likely to ever fill all 1MiB of memory, having a sparse list with
a default value should make it much faster.

As well, added contracts and guards to the sparse list code as a
learning exercise and to fail early.

Lastly, moved all of the checklist from `mips-vm.rkt` into `TODO.md`
to make the code more readable.